+++ src/Entity/TermTag.php
@@ -23,6 +23,12 @@ class TermTag
     #[ORM\Column(name: 'TgComment', length: 200)]
     private ?string $comment = '';
 
+    #[ORM\JoinTable(name: 'wordtags')]
+    #[ORM\JoinColumn(name: 'WtTgID', referencedColumnName: 'TgID')]
+    #[ORM\InverseJoinColumn(name: 'WtWoID', referencedColumnName: 'WoID')]
+    #[ORM\ManyToMany(targetEntity: Term::class, mappedBy:'termTags', cascade: ['persist'])]
+    private Collection $Terms;
+
     public function getId(): ?int
     {
         return $this->id;