+++ src/Repository/BookRepository.php
@@ -29,12 +29,6 @@ class BookRepository extends ServiceEntityRepository
         $stmt->executeQuery();
     }
 
-    private function removeParsedData(int $bookid): void
-    {
-        $this->exec_sql("delete from sentences where SeTxID in (select TxID from texts where TxBkID = $bookid)");
-        $this->exec_sql("delete from texttokens where TokTxID in (select TxID from texts where TxBkID = $bookid)");
-    }
-
     public function save(Book $entity): void
     {
         $isnew = ($entity->getID() == null);
@@ -52,8 +46,17 @@ class BookRepository extends ServiceEntityRepository
 
     public function remove(Book $entity, bool $flush = false): void
     {
+        // THIS IS REQUIRED.  Without the pragma, the deletes don't
+        // propagate throughout the database foreign key cascade
+        // deletes.
+        //
+        // NOTE it may be better to generalize this, per
+        // https://stackoverflow.com/questions/22924444/,
+        //   foreign-key-constraints-does-not-work-with-doctrine-symfony-and-sqlite
+        //   (famas23's answer).
+        $this->getEntityManager()->getConnection()->exec('PRAGMA foreign_keys = ON');
+
         $this->getEntityManager()->remove($entity);
-        $this->removeParsedData($entity->getId());
 
         if ($flush) {
             $this->getEntityManager()->flush();
@@ -69,7 +72,10 @@ class BookRepository extends ServiceEntityRepository
         $base_sql = "SELECT
           b.BkID As BkID,
           LgName,
-          COALESCE(currtext.TxTitle, BkTitle) as BkTitle,
+          BkTitle || case
+            when currtext.TxID is null then ''
+            else ' (' || currtext.TxOrder || '/' || pagecnt.c || ')'
+          end as BkTitle,
           pagecnt.c as PageCount,
           BkArchived,
           tags.taglist AS TagList,