+++ tests/acceptance/AcceptanceTestBase.php
@@ -20,6 +20,7 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\Panther\PantherTestCase;
 
 use App\Entity\Language;
+use App\Entity\Book;
 use App\Entity\Text;
 use App\Entity\Term;
 use App\Entity\TextTag;
@@ -34,10 +35,10 @@ use App\Repository\BookRepository;
 use App\Repository\ReadingRepository;
 use App\Repository\SettingsRepository;
 use App\Domain\TermService;
-use App\Domain\BookBinder;
 use App\Domain\ReadingFacade;
 
 use App\Tests\acceptance\Contexts\ReadingContext;
+use App\Tests\acceptance\Contexts\BookContext;
 use App\Tests\acceptance\Contexts\TermContext;
 use App\Tests\acceptance\Contexts\TermTagContext;
 
@@ -158,19 +159,8 @@ abstract class AcceptanceTestBase extends PantherTestCase
         return $ret;
     }
 
-    public function load_french_data(): void
-    {
-        $this->addTerms($this->french, ['lista']);
-        $frid = $this->french->getLgID();
-        $frt = new Text();
-        $frt->setTitle("Bonjour.");
-        $frt->setText("Bonjour je suis lista.");
-        $frt->setLanguage($this->french);
-        $this->text_repo->save($frt, true);
-    }
-
     public function make_text(string $title, string $text, Language $lang): Text {
-        $b = BookBinder::makeBook($title, $lang, $text);
+        $b = Book::makeBook($title, $lang, $text);
         $this->book_repo->save($b, true);
         $b->fullParse();  // Most tests require full parsing.
         return $b->getTexts()[0];
@@ -240,6 +230,10 @@ abstract class AcceptanceTestBase extends PantherTestCase
 
     public function getReadingContext() {
         return new ReadingContext($this->client);
+    }
+
+    public function getBookContext() {
+        return new BookContext($this->client);
     }
 
     public function getTermContext() {